+++ src/Glpi/Application/SystemConfigurator.php
@@ -39,6 +39,7 @@ use Glpi\Log\AccessLogHandler;
 use Glpi\Log\ErrorLogHandler;
 use Monolog\Logger;
 use Psr\Log\LoggerInterface;
+use Psr\Log\LogLevel;
 
 final class SystemConfigurator
 {
@@ -50,6 +51,9 @@ final class SystemConfigurator
         $this->setSessionConfiguration();
         $this->initLogger();
         $this->registerErrorHandler();
+
+        // Keep it after `registerErrorHandler()` call to be sure that messages are correctly handled.
+        $this->checkForObsoleteConstants();
     }
 
     public function getLogger(): LoggerInterface
@@ -99,7 +103,7 @@ final class SystemConfigurator
 
                 // Where to load plugins.
                 // Order in this array is important (priority to first found).
-                'PLUGINS_DIRECTORIES'  => [
+                'GLPI_PLUGINS_DIRECTORIES' => [
                     '{GLPI_MARKETPLACE_DIR}',
                     $this->root_dir . '/plugins',
                 ],
@@ -145,12 +149,15 @@ final class SystemConfigurator
 
                 // Constants dedicated to developers
                 'GLPI_DISABLE_ONLY_FULL_GROUP_BY_SQL_MODE' => '1', // '1' to disable ONLY_FULL_GROUP_BY 'sql_mode'
+                'GLPI_LOG_LVL'                             => LogLevel::WARNING,
+                'GLPI_SKIP_UPDATES'                        => false, // `true` to bypass minor versions DB updates
                 'GLPI_STRICT_ENV'                          => false, // `true` to make environment more strict (strict variables in twig templates, etc)
 
                 // Other constants
                 'GLPI_AJAX_DASHBOARD'         => '1', // 1 for "multi ajax mode" 0 for "single ajax mode" (see Glpi\Dashboard\Grid::getCards)
                 'GLPI_CALDAV_IMPORT_STATE'    => 0, // external events created from a caldav client will take this state by default (0 = Planning::INFO)
                 'GLPI_CENTRAL_WARNINGS'       => '1', // display (1), or not (0), warnings on GLPI Central page
+                'GLPI_SYSTEM_CRON'            => false, // `true` to use the system cron provided by the downstream package
                 'GLPI_TEXT_MAXSIZE'           => '4000', // character threshold for displaying read more button
                 'GLPI_WEBHOOK_ALLOW_RESPONSE_SAVING' => '0', // allow (1) or not (0) to save webhook response in database
             ],
@@ -177,6 +184,11 @@ final class SystemConfigurator
         // Define constants values from downstream distribution file
         if (!defined('TU_USER') && file_exists($this->root_dir . '/inc/downstream.php')) {
             include_once($this->root_dir . '/inc/downstream.php');
+        }
+
+        // Handle deprecated/obsolete constants
+        if (defined('PLUGINS_DIRECTORIES') && !defined('GLPI_PLUGINS_DIRECTORIES')) {
+            define('GLPI_PLUGINS_DIRECTORIES', PLUGINS_DIRECTORIES);
         }
 
         // Configure environment type if not defined by user.
@@ -268,5 +280,22 @@ final class SystemConfigurator
     {
         $errorHandler = new ErrorHandler($this->logger);
         $errorHandler::register($errorHandler);
+    }
+
+    private function checkForObsoleteConstants(): void
+    {
+        if (defined('GLPI_USE_CSRF_CHECK')) {
+            trigger_error(
+                'The `GLPI_USE_CSRF_CHECK` constant is now ignored for security reasons.',
+                E_USER_WARNING
+            );
+        }
+
+        if (defined('PLUGINS_DIRECTORIES')) {
+            trigger_error(
+                'The `PLUGINS_DIRECTORIES` constant is deprecated. Use the `GLPI_PLUGINS_DIRECTORIES` constant instead.',
+                E_USER_DEPRECATED
+            );
+        }
     }
 }